+++ build_test/lib/src/in_memory_reader_writer.dart
@@ -2,11 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:build/build.dart';
 // ignore: implementation_imports
 import 'package:build/src/internal.dart';
+import 'package:convert/convert.dart';
+import 'package:crypto/crypto.dart';
 import 'package:glob/glob.dart';
 
 /// An [AssetReader] that records which assets have been read to [assetsRead].
@@ -18,8 +19,8 @@ abstract class RecordingAssetReader implements AssetReader {
 /// assets.
 ///
 /// TODO(davidmorgan): merge into `FileBasedReader` and `FileBasedWriter`.
-class InMemoryAssetReaderWriter extends AssetReader
-    implements InMemoryAssetReader, InMemoryAssetWriter {
+class InMemoryAssetReaderWriter
+    implements AssetReader, AssetReaderState, AssetWriter {
   @override
   late final AssetFinder assetFinder = FunctionAssetFinder(_findAssets);
 
@@ -41,27 +42,26 @@ class InMemoryAssetReaderWriter extends AssetReader
   @override
   Filesystem get filesystem => _filesystem;
 
-  @override
   Map<AssetId, List<int>> get assets => _filesystem.assets;
 
   @override
   Future<bool> canRead(AssetId id) async {
     inputTracker.assetsRead.add(id);
-    return assets.containsKey(id);
+    return filesystem.exists(id);
   }
 
   @override
   Future<List<int>> readAsBytes(AssetId id) async {
     if (!await canRead(id)) throw AssetNotFoundException(id);
     inputTracker.assetsRead.add(id);
-    return assets[id]!;
+    return filesystem.readAsBytes(id);
   }
 
   @override
   Future<String> readAsString(AssetId id, {Encoding encoding = utf8}) async {
     if (!await canRead(id)) throw AssetNotFoundException(id);
     inputTracker.assetsRead.add(id);
-    return encoding.decode(assets[id]!);
+    return filesystem.readAsString(id, encoding: encoding);
   }
 
   // This is only for generators, so only `BuildStep` needs to implement it.
@@ -80,37 +80,21 @@ class InMemoryAssetReaderWriter extends AssetReader
   }
 
   @override
-  Future writeAsBytes(AssetId id, List<int> bytes) async {
-    assets[id] = bytes is Uint8List ? bytes : Uint8List.fromList(bytes);
-  }
+  Future writeAsBytes(AssetId id, List<int> bytes) async =>
+      filesystem.writeAsBytes(id, bytes);
 
   @override
   Future writeAsString(AssetId id, String contents,
-      {Encoding encoding = utf8}) async {
-    assets[id] = encoding.encode(contents) as Uint8List;
+          {Encoding encoding = utf8}) async =>
+      filesystem.writeAsString(id, contents, encoding: encoding);
+
+  @override
+  Future<Digest> digest(AssetId id) async {
+    var digestSink = AccumulatorSink<Digest>();
+    md5.startChunkedConversion(digestSink)
+      ..add(await readAsBytes(id))
+      ..add(id.toString().codeUnits)
+      ..close();
+    return digestSink.events.first;
   }
-}
-
-/// An implementation of [AssetReader] with primed in-memory assets.
-abstract class InMemoryAssetReader implements AssetReader, AssetReaderState {
-  abstract final Map<AssetId, List<int>> assets;
-
-  /// Create a new asset reader.
-  ///
-  /// May optionally define a [rootPackage], which is required for some APIs.
-  factory InMemoryAssetReader({String? rootPackage}) =>
-      InMemoryAssetReaderWriter(rootPackage: rootPackage);
-}
-
-/// An implementation of [AssetWriter] that writes outputs to memory.
-abstract class InMemoryAssetWriter implements AssetWriter {
-  abstract final Map<AssetId, List<int>> assets;
-
-  factory InMemoryAssetWriter() => InMemoryAssetReaderWriter();
-
-  @override
-  Future writeAsBytes(AssetId id, List<int> bytes);
-
-  @override
-  Future writeAsString(AssetId id, String contents, {Encoding encoding = utf8});
 }