+++ build_test/test/check_outputs_test.dart
@@ -7,38 +7,38 @@ void main() {
       var a = makeAssetId('a|lib/a.txt');
       var b = makeAssetId('a|lib/b.txt');
       var actualAssets = [a, b];
-      var writer = InMemoryAssetWriter();
-      await writer.writeAsString(a, 'a');
-      await writer.writeAsString(b, 'b');
+      var readerWriter = InMemoryAssetReaderWriter();
+      await readerWriter.writeAsString(a, 'a');
+      await readerWriter.writeAsString(b, 'b');
 
       var outputs = {'a|lib/a.txt': 'a', 'a|lib/b.txt': 'b'};
 
-      expect(
-          () => checkOutputs(outputs, actualAssets, writer), returnsNormally);
+      expect(() => checkOutputs(outputs, actualAssets, readerWriter),
+          returnsNormally);
     });
     test('with extra output', () async {
       var a = makeAssetId('a|lib/a.txt');
       var b = makeAssetId('a|lib/b.txt');
       var actualAssets = [a, b];
-      var writer = InMemoryAssetWriter();
-      await writer.writeAsString(a, 'a');
-      await writer.writeAsString(b, 'b');
+      var readerWriter = InMemoryAssetReaderWriter();
+      await readerWriter.writeAsString(a, 'a');
+      await readerWriter.writeAsString(b, 'b');
 
       var outputs = {'a|lib/a.txt': 'a'};
 
-      expect(() => checkOutputs(outputs, actualAssets, writer),
+      expect(() => checkOutputs(outputs, actualAssets, readerWriter),
           throwsA(const TypeMatcher<TestFailure>()));
     });
 
     test('with missing output', () async {
       var a = makeAssetId('a|lib/a.txt');
       var actualAssets = [a];
-      var writer = InMemoryAssetWriter();
-      await writer.writeAsString(a, 'a');
+      var readerWriter = InMemoryAssetReaderWriter();
+      await readerWriter.writeAsString(a, 'a');
 
       var outputs = {'a|lib/a.txt': 'a', 'a|lib/b.txt': 'b'};
 
-      expect(() => checkOutputs(outputs, actualAssets, writer),
+      expect(() => checkOutputs(outputs, actualAssets, readerWriter),
           throwsA(const TypeMatcher<TestFailure>()));
     });
 
@@ -47,14 +47,14 @@ void main() {
       var b = makeAssetId('b|lib/b.txt');
       var bMapped = makeAssetId('a|.generated/b/lib/b.txt');
       var actualAssets = [a, b];
-      var writer = InMemoryAssetWriter();
-      await writer.writeAsString(a, 'a');
-      await writer.writeAsString(bMapped, 'b');
+      var readerWriter = InMemoryAssetReaderWriter();
+      await readerWriter.writeAsString(a, 'a');
+      await readerWriter.writeAsString(bMapped, 'b');
 
       var outputs = {'a|lib/a.txt': 'a', 'b|lib/b.txt': 'b'};
 
       expect(
-          () => checkOutputs(outputs, actualAssets, writer,
+          () => checkOutputs(outputs, actualAssets, readerWriter,
               mapAssetIds: (id) => id == b ? bMapped : a),
           returnsNormally);
     });